test(progressView): consolidate LogList coverage
