+++ src/test-kernel/progressView/LogListKeyboard.vitest.mts
@@ -14,6 +14,12 @@ useLitComponentTestDom(
 interface LogListTestHooks {
   handleKeyEvent(event: Event): void;
   activateLinkFromEvent(event: Event): boolean;
+  streamCache: {
+    readonly size: number;
+    has(streamId: string): boolean;
+    rkeys(): Generator<string, void, unknown>;
+  };
+  getOrCreateEntry(streamId: string): unknown;
 }
 
 function setupEnterKeyOnLink(): {
@@ -60,5 +66,49 @@ describe('log-list keyboard activation', () => {
     hooks.handleKeyEvent(event);
 
     expect(activateLinkFromEvent).not.toHaveBeenCalled();
+  });
+});
+
+describe('log-list stream cache', () => {
+  function createCacheHooks(): LogListTestHooks {
+    return document.createElement('log-list') as unknown as LogListTestHooks;
+  }
+
+  it('reuses cached entries and promotes them to most recently used', () => {
+    const hooks = createCacheHooks();
+    const originalEntry = hooks.getOrCreateEntry('stream-1');
+
+    hooks.getOrCreateEntry('stream-2');
+    const reusedEntry = hooks.getOrCreateEntry('stream-1');
+
+    expect(reusedEntry).toBe(originalEntry);
+    expect([...hooks.streamCache.rkeys()]).toEqual(['stream-2', 'stream-1']);
+  });
+
+  it('evicts the least-recently used stream at the five-entry bound', () => {
+    const hooks = createCacheHooks();
+
+    for (const streamId of [
+      'stream-1',
+      'stream-2',
+      'stream-3',
+      'stream-4',
+      'stream-5',
+    ]) {
+      hooks.getOrCreateEntry(streamId);
+    }
+
+    hooks.getOrCreateEntry('stream-6');
+
+    expect(hooks.streamCache.size).toBe(5);
+    expect(hooks.streamCache.has('stream-1')).toBe(false);
+    expect(hooks.streamCache.has('stream-2')).toBe(true);
+    expect([...hooks.streamCache.rkeys()]).toEqual([
+      'stream-2',
+      'stream-3',
+      'stream-4',
+      'stream-5',
+      'stream-6',
+    ]);
   });
 });
